refactor : InGameUserRow.types.ts로 types 이동 및 폴더를 components로 이동
