+++ components/inGameUserRow/InGameUserRow.types.ts
@@ -1,0 +1,8 @@
+import { RuneInfo, SpellInfos } from 'types';
+
+export interface InGameUserRowProps {
+  nickname: string;
+  runes: RuneInfo[];
+  spells: SpellInfos;
+  profileIconId: number;
+}
